Move order_status_change function to signals file
